pyproject.toml
[tool.poetry]
name = "datapane"
version = "0.0.1"
description = "Datapane client library and CLI tool"
authors = ["Datapane Team <dev@datapane.com>"]
license = "Proprietary"
readme = "README.md"
homepage = "https://www.datapane.com"
repository = "https://www.github/datapane/datapane-cli"
documentation = "https://docs.datapane.com"


keywords = ["datapane", "data", "jupyter", "pandas", "altair"]
classifiers = [
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Environment :: Console",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    # "Private :: Do Not Upload",  # hack to force invalid package for upload
]

#packages = [
#    { include = "datapane", from = "src" }
#]

# extra files (allows overriding gitignore)
include = ["README.md", "LICENSE.txt"]

[tool.poetry.dependencies]
python = "^3.6.1"
dataclasses = {version = "0.7", python = "~3.6.1"}

# DS
numpy = "^1.18.0"
pandas = "^1.0.1"
scikit-learn = "^0.22.1"
pyarrow = "0.17.*"

# viz
matplotlib = "^3.1.0"
altair = ">=3.0.1"
bokeh = "^2.0.0"
nbconvert = "^5.6.1"
# ipython = "^7.7.0" # needed for nbconvert?

# misc
dacite = "^1.2.0"
colorlog = "^4.1.0"
munch = "^2.5.0"
stringcase = "^1.2.0"
PyYAML = "^5.3.0"
packaging = "^20.3"

# CLI
click = "^7.0.0"
click-spinner = "^0.1.8"
# httpx = "^0.7.5"
"ruamel.yaml" = "^0.16.5"
requests = "^2.20.0"
validators = "^0.15.0"
jinja2 = "^2.11.1"
jsonschema = "^3.2.0"
tabulate = "^0.8.7"
# TODO - check for appdir releases
# appdirs = "^1.4.3"
toolz = "^0.10.0"
flit-core = "~2.3.0"
importlib_resources = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-datadir = "^1.3.1"
jupyterlab = "^2.1.2"

[tool.poetry.scripts]
datapane = "datapane.client.__main__:main"
dp-runner = "datapane.runner.__main__:main"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
